[dynamic-shapes] Split dynamic_api_test.py
PiperOrigin-RevId: 461109288
